Learn_model/requirements.txt
ultralytics
opencv-python
numpy
albumentations
tqdm
pyyaml
